+++ .rubocop.yml
@@ -14,7 +14,7 @@ Metrics/MethodLength:
 Metrics/AbcSize:
   Max: 50
 Metrics/ClassLength:
-  Max: 150
+  Max: 200
 Metrics/BlockLength:
   IgnoredMethods: ['describe']
   Max: 30
