+++ classes/books.json
@@ -18,5 +18,9 @@
   {
     "title": "t1",
     "author": "a1"
+  },
+  {
+    "title": "666",
+    "author": "45"
   }
 ]
