+++ classes/rentals.json
@@ -128,5 +128,18 @@
       "age": 18,
       "name": "ben"
     }
+  },
+  {
+    "date": "12-12-2023",
+    "book": {
+      "title": "666",
+      "author": "45"
+    },
+    "person": {
+      "type": "teacher",
+      "id": 79,
+      "age": 666,
+      "name": "666"
+    }
   }
 ]
